# backend/swagger.yaml
# openapi: 3.0.0
# info:
#   title: Recently Viewed Products API
#   version: 1.0.0
# paths:
#   /api/v1/users/{userId}/recentlyViewed:
#     get:
#       summary: Get recently viewed products
#       parameters:
#         - name: userId
#           in: path
#           required: true
#           schema:
#             type: string
#       security:
#         - BearerAuth: []
#       responses:
#         200:
#           description: List of recently viewed products
#         401:
#           description: Unauthorized
#         500:
#           description: Server error
# components:
#   securitySchemes:
#     BearerAuth:
#       type: http
#       scheme: bearer


openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
  description: API documentation for your application.
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /api/v1/users/{userId}/recentlyViewed:
    get:
      summary: Get recently viewed products by a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: ID of the user
      responses:
        '200':
          description: Successfully fetched recently viewed products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    productId:
                      type: string
                    name:
                      type: string
                    timestamp:
                      type: string
                      format: date-time
                    viewCount:
                      type: integer
        '500':
          description: Failed to fetch recently viewed products
    post:
      summary: Add a recently viewed product for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: ID of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
      responses:
        '200':
          description: Product view recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  productId:
                    type: string
                  userId:
                    type: string
        '404':
          description: Product not found
        '500':
          description: Failed to record product view
  /api/v1/users/products:
    get:
      summary: Get all products
      responses:
        '200':
          description: Successfully fetched products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    description:
                      type: string
        '500':
          description: Failed to fetch products
  /api/v1/users/products/{productId}:
    get:
      summary: Get product details by ID
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
          description: ID of the product
      responses:
        '200':
          description: Successfully fetched product details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  description:
                    type: string
        '404':
          description: Product not found
        '500':
          description: Failed to fetch product details